Numerous updates to code and tests
Fixed a major bug where dynamic browse was using some logic from static browse, due to a change in input files.

Revised the way that PropBrowseInReach is calculated to better represent logic as laid out in the user guide. This change required adding a new input variable, BrowseBiomassThresholdMin, which dictates the biomass below which a cohort can be entirely browsed. PropBrowseInReach now follows a sensible functional relationship, where larger cohorts are less susceptible to browse.

Attempted to fix an error where Browse Proportion is greater than 1. Sometimes this error is caused by converting to integers, which gives values that are essentially equal to 1 but still don't meet the test. However, some cohorts are still giving values of Browse Proportion around 1.2 - 1.5, so something else needs to be fixed here.

Set ANPP = Biomass * 0.04 instead of 0.10, which much improved estimates of cohort forage. Still a placeholder, but this matches average values from Biomass Succession fairly well.

+++ src/EventsLog.cs
@@ -22,7 +22,7 @@ namespace Landis.Extension.Browse
         public int TotalPopulation { get; set; }
 
         [DataFieldAttribute(Desc = "Population Density (animals per ha)")]
-        public int PopulationDensity { get; set; }
+        public double PopulationDensity { get; set; }
 
         [DataFieldAttribute(Desc = "Total Forage (kg)")]
         public int TotalForage { get; set; }
